refactor: modularize apts/objects into a package
- Split the 593-line 'God Object' apts/objects/objects.py into base.py, utils.py, and subclasses.
- Extracted refraction and vectorized geometric calculations into apts/objects/utils.py.
- Moved star-specific optimizations to apts/objects/stars.py.
- Renamed _fast_compute_stars to _vectorized_geometric_compute for package-wide consistency.
- Maintained public API compatibility via __init__.py re-exports.
- Updated all internal and external import references.
- Verified 100% test pass rate (663/663).
