+++ apts/objects/__init__.py
@@ -1,6 +1,7 @@
+from .base import Objects as Objects
 from .messier import Messier as Messier
 from .solar_objects import SolarObjects as SolarObjects
 from .ngc import NGC as NGC
 from .stars import Stars as Stars
 
-__all__ = ["Messier", "SolarObjects", "NGC", "Stars"]
+__all__ = ["Objects", "Messier", "SolarObjects", "NGC", "Stars"]
