+++ apts/objects/base.py
@@ -7,13 +7,13 @@ from typing import Any, cast
 
 import numpy as np
 import pandas as pd
-import pytz
 from skyfield import almanac
 from skyfield.api import Star
 from skyfield.searchlib import find_discrete
 
 from ..cache import get_timescale
 from ..constants import ObjectTableLabels
+from .utils import calculate_refraction, vectorized_geometric_compute
 
 logger = logging.getLogger(__name__)
 
@@ -176,7 +176,7 @@ class Objects(ABC):
                 # max_alt = 90 - abs(lat - dec)
                 max_alt_deg = 90.0 - np.abs(self.place.lat_decimal - stars_decs)
                 # Add refraction at max altitude for better filtering
-                max_alt_deg += self._refraction(max_alt_deg)
+                max_alt_deg += calculate_refraction(max_alt_deg)
 
                 # Only process stars that can potentially reach the minimum altitude
                 potential_mask = max_alt_deg > conditions.horizon.get_min_altitude()
@@ -221,7 +221,7 @@ class Objects(ABC):
 
                     # Convert to degrees and add atmospheric refraction
                     true_alt_deg = np.rad2deg(np.arcsin(np.clip(sin_alt, -1.0, 1.0)))
-                    apparent_alt_deg = true_alt_deg + self._refraction(true_alt_deg)
+                    apparent_alt_deg = true_alt_deg + calculate_refraction(true_alt_deg)
 
                     # Determine azimuth
                     # tan(Az) = sin(H) / (cos(H) sin(lat) - tan(dec) cos(lat))
@@ -315,69 +315,12 @@ class Objects(ABC):
         # Create body at given coordinates
         return Star(ra_hours=RA, dec_degrees=Dec)
 
-    @staticmethod
-    def _refraction(alt_deg):
-        """
-        Calculates atmospheric refraction in degrees using Bennett's formula.
-        Accurate to ~0.02' for altitudes > 0.
-        """
-        alt_deg_arr = np.atleast_1d(alt_deg)
-        refraction_deg = np.zeros_like(alt_deg_arr, dtype=float)
-        # Apply only for objects above -1 degree to avoid tan(0) or instability
-        mask = alt_deg_arr > -1.0
-        if np.any(mask):
-            # R in arcminutes = 1 / tan(h + 7.31 / (h + 4.4))
-            r_arcmin = 1.0 / np.tan(
-                np.deg2rad(alt_deg_arr[mask] + 7.31 / (alt_deg_arr[mask] + 4.4))
-            )
-            refraction_deg[mask] = r_arcmin / 60.0
-
-        return refraction_deg[0] if np.isscalar(alt_deg) else refraction_deg
-
     def _compute_tranzit(self, skyfield_object, observer):
         """
         Calculates the upper meridian transit of a celestial object.
-        For stars, a fast sidereal time approximation is used.
         """
         if skyfield_object is None:
             return None
-
-        # Optimization for stars: use sidereal time formula
-        if isinstance(skyfield_object, Star):
-            current_dt = observer.date.utc_datetime()
-            # Start search from the beginning of the UTC day
-            t0_dt = current_dt.replace(
-                hour=0, minute=0, second=0, microsecond=0, tzinfo=pytz.UTC
-            )
-            t0 = self.ts.utc(t0_dt)
-
-            # RA of the star
-            ra_hours = skyfield_object.ra.hours
-            lon_hours = self.place.lon_decimal / 15.0
-
-            # LST = GMST + lon
-            # We want LST == RA => GMST + lon == RA => GMST == RA - lon
-            target_gmst = (ra_hours - lon_hours) % 24
-
-            current_gmst = t0.gmst
-
-            # Sidereal day is shorter than solar day
-            # 1 solar hour = 1.0027379 sidereal hours
-            # 1 sidereal hour = 0.99726957 solar hours
-            sidereal_to_solar = 0.99726957
-
-            dt_sidereal = (target_gmst - current_gmst) % 24
-            dt_solar = dt_sidereal * sidereal_to_solar
-
-            transit_dt = t0_dt + timedelta(hours=dt_solar)
-
-            # Ensure we catch the transit relevant to the observation window
-            # If it already happened more than 12 hours before observation start,
-            # we might want the next one (stars transit once per sidereal day)
-            if transit_dt < current_dt - timedelta(hours=12):
-                transit_dt += timedelta(hours=24 * sidereal_to_solar)
-
-            return transit_dt.astimezone(observer.local_timezone)
 
         # Fallback for moving objects (planets)
         current_dt = observer.date.utc_datetime()
@@ -409,7 +352,6 @@ class Objects(ABC):
     def _compute_rising_and_setting(self, skyfield_object, observer, transit_time):
         """
         Calculates rising and setting times for a celestial object.
-        Note: For stars, use _fast_compute_stars for better performance.
         """
         if skyfield_object is None or transit_time is None or pd.isna(transit_time):
             return None, None
@@ -454,15 +396,6 @@ class Objects(ABC):
         if transit is None or pd.isna(transit):
             return 0
 
-        # Optimization for stars: geometric formula
-        if isinstance(skyfield_object, Star):
-            lat = self.place.lat_decimal
-            dec = skyfield_object.dec.degrees
-            # Max altitude = 90 - abs(lat - dec)
-            # Add refraction for better accuracy
-            true_alt = 90.0 - abs(lat - dec)
-            return true_alt + self._refraction(true_alt)
-
         t = self.ts.utc(transit)
         alt, _, _ = (
             self.place.observer.at(t)
@@ -472,7 +405,7 @@ class Objects(ABC):
         )
         return alt.degrees
 
-    def _fast_compute_stars(self, df, observer):
+    def _vectorized_geometric_compute(self, df, observer):
         """
         Fast transit, altitude, rising, and setting calculation for a DataFrame of Stars.
         Uses vectorized numpy operations and geometric approximations for speed.
@@ -498,96 +431,14 @@ class Objects(ABC):
             [obj.dec.degrees if isinstance(obj, Star) else 0 for obj in sky_objs]
         )
 
-        return self._vectorized_geometric_compute(df, observer, ras, decs, valid_mask)
-
-    def _vectorized_geometric_compute(self, df, observer, ras, decs, valid_mask):
-        """
-        Generic vectorized transit, altitude, rising, and setting calculation.
-        Uses vectorized numpy operations and geometric approximations for speed.
-
-        Note: Rising/setting times use a geometric formula that accounts for atmospheric
-        refraction (~34'). This results in high accuracy compared to Skyfield's
-        iterative solver, while maintaining excellent performance.
-        """
-        current_dt = observer.date.utc_datetime()
-        t0_dt = current_dt.replace(
-            hour=0, minute=0, second=0, microsecond=0, tzinfo=pytz.UTC
+        return vectorized_geometric_compute(
+            self.ts,
+            self.place.lat_decimal,
+            self.place.lon_decimal,
+            observer.local_timezone,
+            observer.date,
+            ras,
+            decs,
+            valid_mask,
+            len(df),
         )
-        t0 = self.ts.utc(t0_dt)
-        lon_hours = self.place.lon_decimal / 15.0
-        current_gmst = t0.gmst
-        sidereal_to_solar = 0.99726957
-        lat_deg = self.place.lat_decimal
-
-        # Vectorized Transit calculation
-        target_gmst = (ras - lon_hours) % 24
-        dt_solar = ((target_gmst - current_gmst) % 24) * sidereal_to_solar
-
-        # Use pandas for datetime vectorization
-        t0_ts = pd.Timestamp(t0_dt)
-        # Ensure second precision to avoid lossless cast errors in newer pandas
-        transit_times = (t0_ts + pd.to_timedelta(dt_solar * 3600, unit="s")).floor("s")
-
-        # Adjust for 12-hour window relative to current time
-        cutoff = current_dt - timedelta(hours=12)
-        shift = timedelta(hours=24 * sidereal_to_solar)
-
-        transit_times = pd.Series(transit_times)
-        # Ensure cutoff is compatible with transit_times timezone
-        cutoff_ts = pd.Timestamp(cutoff)
-        if transit_times.dt.tz is None and cutoff_ts.tz is not None:
-            cutoff_ts = cutoff_ts.replace(tzinfo=None)
-        elif transit_times.dt.tz is not None and cutoff_ts.tz is None:
-            cutoff_ts = cutoff_ts.replace(tzinfo=pytz.UTC)
-
-        needs_shift = transit_times < cutoff_ts
-        transit_times.loc[needs_shift] += shift
-
-        # Localize transits
-        # Optimization: Bulk timezone conversion is ~13x faster than iterative .astimezone()
-        local_tz = observer.local_timezone
-        transit_times_local = transit_times.dt.tz_convert(local_tz).to_numpy(dtype=object)
-        transits = [
-            t if m and pd.notna(t) else None
-            for t, m in zip(transit_times_local, valid_mask)
-        ]
-
-        # Vectorized Altitude calculation
-        altitudes = 90.0 - np.abs(lat_deg - decs)
-        # Add atmospheric refraction for high-precision
-        altitudes += self._refraction(altitudes)
-        alts = [float(a) if m else 0 for a, m in zip(altitudes, valid_mask)]
-
-        # Vectorized Rise/Set (Geometric) calculation
-        lat_rad = np.deg2rad(lat_deg)
-        decs_rad = np.deg2rad(decs)
-        # Standard altitude for rising/setting of stars is -34 arcminutes to account for refraction
-        h0_rad = np.deg2rad(-34.0 / 60.0)
-        # cos(H) = (sin(h0) - sin(lat)sin(dec)) / (cos(lat)cos(dec))
-        cos_H = (np.sin(h0_rad) - np.sin(lat_rad) * np.sin(decs_rad)) / (
-            np.cos(lat_rad) * np.cos(decs_rad)
-        )
-
-        # Hour angle in solar hours
-        H_hours = np.full(len(df), np.nan)
-        # Objects must be Stars and within the range where they actually rise/set
-        h_mask = valid_mask & (cos_H >= -1) & (cos_H <= 1)
-        H_hours[h_mask] = (
-            np.arccos(np.clip(cos_H[h_mask], -1.0, 1.0))
-            * (12.0 / np.pi)
-            * sidereal_to_solar
-        )
-
-        # Ensure second precision
-        H_delta = cast(Any, pd.to_timedelta(H_hours * 3600, unit="s")).round("s")
-        rising_times = (transit_times - H_delta).dt.floor("s")
-        setting_times = (transit_times + H_delta).dt.floor("s")
-
-        # Optimization: Bulk timezone conversion for rise/set times
-        rising_times_local = rising_times.dt.tz_convert(local_tz).to_numpy(dtype=object)
-        setting_times_local = setting_times.dt.tz_convert(local_tz).to_numpy(dtype=object)
-
-        rises = [t if pd.notna(t) else None for t in rising_times_local]
-        sets = [t if pd.notna(t) else None for t in setting_times_local]
-
-        return transits, alts, rises, sets
