+++ apts/objects/ngc.py
@@ -2,7 +2,7 @@ from types import SimpleNamespace
 
 import pandas as pd
 
-from .objects import Objects
+from .base import Objects
 from ..catalogs import Catalogs
 from ..constants import ObjectTableLabels
 
@@ -155,7 +155,7 @@ class NGC(Objects):
         computed_df = df_to_compute.copy()
 
         # Fast transit and altitude calculation for stars
-        transits, alts, rises, sets = self._fast_compute_stars(
+        transits, alts, rises, sets = self._vectorized_geometric_compute(
             computed_df, observer_to_use
         )
         computed_df[ObjectTableLabels.TRANSIT] = transits
