+++ requirements_no_cairo.txt
@@ -1,0 +1,18 @@
+aenum==3.1.16
+Babel
+ephem==4.2
+matplotlib==3.10.8
+numpy==2.3.5
+pandas==2.3.3
+Pint==0.25.2
+networkx==3.6.1
+pytz==2025.2
+requests-cache==1.2.1
+requests-mock==1.12.1
+scipy==1.16.3
+seaborn==0.13.2
+skyfield==1.54
+svgwrite==1.4.3
+timezonefinder==8.1.0
+astropy==7.2.0
+photutils==2.3.0
